Collider drawing
Collider now scales and offsets correctly. Next step is to handle this nicely in C++. Will make an addSpriteCollider function that takes the details from a sprite and creates the component by default

+++ Engine/src/Scripting/Fysics/FysicsScripting.cpp
@@ -166,8 +166,8 @@ namespace IonixEngine
 
 		//----------Collision Methods----------
 
-		auto addBoxCollider = [](Entity* entity, float sizeX, float sizeY, int offsetX, int offsetY, float angle, bool isTrigger) {
-
+		auto addBoxCollider = [](Entity* entity, float sizeX, float sizeY, float offsetX, float offsetY, float angle, bool isTrigger) {
+			
 			b2Vec2 size;
 			size.x = sizeX;
 			size.y = sizeY;
